Fix the idiosyncratic "no atoms??" error message.
This Lua error appears if the atoms argument to pd.Class:outlet is
anything but a table. The fix is trivial: Check that the atoms argument
actually is a table, and, if this is not the case, either output a more
helpful error message, or just turn the singleton into a table with one
element (which is what we do here).

I'm not sure that I actually want to fix this. It's one of pdlua's
lovely quirks which I actually discuss in the tutorial in order to
illustrate what happens if something goes wrong deep down in pdlua's
internals.