Revert "Merge dofile and doclassfile back into one."
We really need to keep these separate since the externdir of a Lua class
isn't necessarily the same as the canvas dir containing the object any
more. The former would be used, e.g., in cases where you want to reload
a class file for live coding purposes, whereas the latter would be used
to load Lua code relative to the object's canvas dir (in particular,
that would be the case for anonymous classes created using pdluax).

This reverts commit f2a889c6ab822f9049ca5bba964c675a5f7e127e.