Work around fallout from rev. a4621af which breaks the dofile mechanism.
Specifically, those trailing slashes on pd._loadpath in
pd.Class:register keep piling up, which means that an existing class
will not be found when being invoked from dofile/dofilex. We work around
this by checking for a trailing slash now. Hopefully, this is only a
temporary kludge until a proper fix is deployed.