Fix iolets being drawn on the wrong canvas.
glist_drawiofor and glist_eraseiofor were being invoked on the wrong
glist argument (i.e., in the parent of the glist instead of the glist
itself). This probably didn't affect vanilla much, but in purr-data we
rely on that data being accurate.

Also made sure in the purr-data implementation that we actually check
for canvas != glist to avoid drawing the iolets in a subpatch, if a
graphics object lives in the gop area of a subpatch. That's the way that
all the other objects in purr-data behave inside a gop area.

None of this will affect the plugdata implementation since all changes
are in the vanilla/purr-data part of the implementation.