Add the pdx.lua support files to the installation.
This makes pdx.lua, which is well-tested at this point and is by far the
easiest live-coding solution for pd-lua, more accessible to the user.

Also make sure that the pdlua external dir is on Lua's package.(c)path
while loading a pd_lua file, so that Lua scripts (like pdx.lua) and Lua
C modules will be found there when required by the pd_lua module.