Fix some corner cases in pdlua_menu_open.
In particular, for pdluax the menu open wouldn't work at all, because
the script name of the object already includes the full path, so
prepending the externdir to it didn't do much good.

We also deal with some rare cases in which the unregistered objects
(pdlua, pdluax, and some other internal classes) would provide a
relative externdir instead of an absolute one. I noticed this, in
particular, while running an uninstalled test build inside the source
tree. This would also result in pdlua_datadir being a relative path.
These cases should be handled correctly now, so that the proper
absolute paths are being used instead.