Reenable standard text widget behavior in non-gui objects.
This was broken ever since the graphics were added. Now we defer setting
the special gui widget behavior until the gui is actually created.

To make this work, we now keep track of the Pd object class inside the
pdlua_t struct on the C side. Going through a Lua call just to retrieve
that class pointer is a waste of time IMHO.

Also, we now redraw the iolets in case the paint method got added during
a reload (pdx.lua), since this will change the position of the iolets
and the connected patch cords.