Simplify pdlua_menu_open some more.
Now that we store the class pointer on the C side, we might as well put
it to good use in pdlua_menu_open, which saves us one lua_pcall.

So the get_class() Lua method isn't actually used in our sources
anymore. But we still keep it around since it has been there pretty much
forever, so 3rd party code might depend on it.