pdx.lua: Refactoring the pdluax receiver callback.
The actual reloading of the script is still executed exactly once, in
the object with the receiver (doing it more than once would be a waste
of time). However, the rest of the pre- and post-reload processing and
bookkeeping is now executed for each object in the class, so that the
status of each object gets updated, in particular gui and iolets.