Added time utilities TIMEUNITPERMSEC, systime, and timesince.
While the clock:set() method has been available forever, it was pretty
useless without a way to compute the absolute systime values required as
an argument. So we now provide the pd.TIMEUNITPERMSEC constant and the
pd.systime() call which actually make clock:set() a useful alternative
to clock:delay().

As a bonus, we also added pd.timesince() which in conjuction with
pd.systime() allows you to calculate time differences in msec, just like
Pd's built-in timer object.