Auto-destruct clocks and receivers with their target object.
They can still be destroyed manually, but this isn't necessary anymore,
as it will happen automatically during object destruction when needed.

This was one of the major pain points of pd-lua for me, which would
actually cause Pd to crash if the user wasn't extremely careful with
these objects. I always considered this a horrible usability issue if
not an outright bug. Thankfully this is now a thing of the past, and I
can happily remove the corresponding caveats from the tutorial.