Fix an uninitialized return value in pdlua_popatomtable.
It seems that, at least with current gcc versions, the *count value,
which wasn't set anywhere, always happened to be zero, causing a
non-table argument value to be interpreted as an empty table. Thus it
failed to trigger an error message (the infamous "no atoms??" error) in
the callers (pdlua_outlet and pdlua_send).

This is now fixed, so "no atoms??" is back, at least for now. ;-)