Gently improve the error messages in pdlua_outlet and pdlua_send.
These have bothered me for a long time. They are cryptic (especially the
"no atoms??" message, but the rest isn't much better) and don't tell the
user anything that might help to locate the source of the error.

So we now add some information about the outlet number and receiver, if
we have it, and tell the user as precisely as possible which argument is
invalid and why. So no "no atoms??" anymore, sorry. ;-)

We also removed the "not a table" message from pdlua_popatomtable
because it doesn't add any useful information, and the error condition
is already being flagged in the callers anyway.