Fix up the pdlua-meta.pd file, and make sure that it gets installed.
Also added a tutorial-meta.pd file for the tutorial which deserves its
own PDDP meta information.

This is mostly for the benefit of Pd flavors supporting the (old) PDDP
standard (Pd-extended and its derivatives). Purr Data, specifically,
will use the descriptions in the meta files in its help browser to
automatically provide helpful descriptions when bookmarking directories
with help patches. (In fact this works with *any* directory dir
containing a dir-meta.pd file with the usual PDDP meta data in it.)