Also add cpath in pdlua_packagepath.
We don't need this right now, as pdx.lua is the only module in the pdlua
external directory that we load, which is Lua source only. However, we
might want to add other extensions that require a C module in the
future, or the user might want to put other Lua packages such as numlua
there for easy access.