set_args: Update the text in the object box.
This makes sure that the object text on the canvas reflects what's been
set with set_args(). Previously the object text was never updated, which
caused a lot of confusion because of the mismatch between visual
representation and internal state.

Note that the object text is invisible for graphical objects, so there
are no visible changes for those, but the object text gets updated in
either case.