At Ben's request, added a get_args method.
self:get_args() returns an object's arguments in a Lua table. float
atoms are returned as numbers, other types of atoms as a string using
their print representation in Pd syntax (including escapes of special
symbols according to Pd's quoting rules).

Note that there are some special kinds of Pd atoms such as pointers
which don't have a serializable string representation. But as long as
all arguments are float and symbol atoms, the resulting table can be
passed to self:set_table() to re-create the same object arguments.